requirements.txt
fastapi
uvicorn
openai
pypdf
python-dotenv
llama-index>=0.9.26
llama-index-vector-stores-pinecone
llama-index-embeddings-openai